move DummyStorage and DummyRedis into utils/testing_utils.py to DRY out the tests
