+++ .coveragerc
@@ -15,4 +15,4 @@ show_missing = True
 omit =
     **/*_test.py
     redis_consumer/pbs/*
-    redis_consumer/grpc_clients.py
+    redis_consumer/testing_utils.py
