+++ redis_consumer/consumers/base_consumer_test.py
@@ -28,142 +28,18 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import copy
+import itertools
 import json
-import os
+import time
 
 import numpy as np
-from skimage.external import tifffile as tiff
 
 import pytest
 
 from redis_consumer import consumers
-from redis_consumer import utils
 from redis_consumer import settings
 
-
-def _get_image(img_h=300, img_w=300):
-    bias = np.random.rand(img_w, img_h) * 64
-    variance = np.random.rand(img_w, img_h) * (255 - 64)
-    img = np.random.rand(img_w, img_h) * variance + bias
-    return img
-
-
-class Bunch(object):
-    def __init__(self, **kwds):
-        self.__dict__.update(kwds)
-
-
-class DummyRedis(object):
-    # pylint: disable=W0613,R0201
-    def __init__(self, items=[], prefix='predict', status='new'):
-        self.work_queue = copy.copy(items)
-        self.processing_queue = []
-        self.prefix = '/'.join(x for x in prefix.split('/') if x)
-        self.status = status
-        self._redis_master = self
-        self.keys = [
-            '{}:{}:{}'.format(self.prefix, 'x.tiff', self.status),
-            '{}:{}:{}'.format(self.prefix, 'x.zip', 'other'),
-            '{}:{}:{}'.format('other', 'x.TIFF', self.status),
-            '{}:{}:{}'.format(self.prefix, 'x.ZIP', self.status),
-            '{}:{}:{}'.format(self.prefix, 'x.tiff', 'other'),
-            '{}:{}:{}'.format('other', 'x.zip', self.status),
-        ]
-
-    def rpoplpush(self, src, dst):
-        if src.startswith('processing'):
-            source = self.processing_queue
-            dest = self.work_queue
-        elif src.startswith(self.prefix):
-            source = self.work_queue
-            dest = self.processing_queue
-
-        try:
-            x = source.pop()
-            dest.insert(0, x)
-            return x
-        except IndexError:
-            return None
-
-    def lpush(self, name, *values):
-        self.work_queue = list(reversed(values)) + self.work_queue
-        return len(self.work_queue)
-
-    def lrem(self, name, count, value):
-        self.processing_queue.remove(value)
-        return count
-
-    def llen(self, queue):
-        if queue.startswith('processing'):
-            return len(self.processing_queue)
-        return len(self.work_queue)
-
-    def hmget(self, rhash, *args):
-        return [self.hget(rhash, a) for a in args]
-
-    def hmset(self, rhash, hvals):
-        return hvals
-
-    def expire(self, name, time):
-        return 1
-
-    def hget(self, rhash, field):
-        if field == 'status':
-            return rhash.split(':')[-1]
-        elif field == 'file_name':
-            return rhash.split(':')[1]
-        elif field == 'input_file_name':
-            return rhash.split(':')[1]
-        elif field == 'output_file_name':
-            return rhash.split(':')[1]
-        elif field == 'reason':
-            return 'reason'
-        return False
-
-    def hset(self, rhash, status, value):
-        return {status: value}
-
-    def hgetall(self, rhash):
-        return {
-            'model_name': 'model',
-            'model_version': '0',
-            'postprocess_function': '',
-            'preprocess_function': '',
-            'file_name': rhash.split(':')[1],
-            'input_file_name': rhash.split(':')[1],
-            'output_file_name': rhash.split(':')[1],
-            'status': rhash.split(':')[-1],
-            'children': 'predict:1.tiff:done,predict:2.tiff:failed,predict:3.tiff:new',
-            'children:done': 'predict:4.tiff:done,predict:5.tiff:done',
-            'children:failed': 'predict:6.tiff:failed,predict:7.tiff:failed',
-        }
-
-
-class DummyStorage(object):
-    # pylint: disable=W0613,R0201
-    def __init__(self, num=3):
-        self.num = num
-
-    def download(self, path, dest):
-        if path.lower().endswith('.zip'):
-            paths = []
-            for i in range(self.num):
-                img = _get_image()
-                base, ext = os.path.splitext(path)
-                _path = '{}{}{}'.format(base, i, ext)
-                tiff.imsave(os.path.join(dest, _path), img)
-                paths.append(_path)
-            return utils.zip_files(paths, dest)
-        img = _get_image()
-        tiff.imsave(os.path.join(dest, path), img)
-        return path
-
-    def upload(self, zip_path, subdir=None):
-        return True, True
-
-    def get_public_url(self, zip_path):
-        return True
+from redis_consumer.testing_utils import Bunch, DummyStorage, redis_client
 
 
 class TestConsumer(object):
